Implement UI and also navigation

+++ composeApp/src/commonMain/kotlin/feature/news/data/api/NewsDetailApiService.kt
@@ -3,5 +3,5 @@ package feature.news.data.api
 import feature.news.data.model.response.ArticleDetailResponse
 
 interface NewsDetailApiService {
-    suspend fun getNewsDetail(id: Int): ArticleDetailResponse
+    suspend fun getNewsDetail(id: Long): ArticleDetailResponse
 }
