+++ composeApp/src/commonMain/kotlin/feature/news/data/api/NewsDetailApiServiceImpl.kt
@@ -11,7 +11,7 @@ class NewsDetailApiServiceImpl(
     private val httpClient: HttpClient,
     private val appConfig: AppConfig
 ) : NewsDetailApiService {
-    override suspend fun getNewsDetail(id: Int): ArticleDetailResponse {
+    override suspend fun getNewsDetail(id: Long): ArticleDetailResponse {
         val xid = id % 8
         val url = "${appConfig.baseApiUrl}details"
         return httpClient.get(url) {
