+++ composeApp/src/commonMain/kotlin/feature/news/data/datasource/remote/NewsDetailRemoteDataSourceImpl.kt
@@ -9,7 +9,7 @@ class NewsDetailRemoteDataSourceImpl(
     private val apiService: NewsDetailApiService
 ) : NewsDetailDataSource.Remote {
 
-    override suspend fun getNewsDetail(id: Int): ArticleDetailResponse {
+    override suspend fun getNewsDetail(id: Long): ArticleDetailResponse {
         return try {
             apiService.getNewsDetail(id)
         } catch (e: Exception) {
